+++ Common/src/main/java/shadowmaster435/impactfulweather/client/particle/WeepingTearSplash.java
@@ -30,7 +30,7 @@ public class WeepingTearSplash extends SimpleAnimatedParticle {
         this.xo = this.x;
         this.yo = this.y;
         this.zo = this.z;
-        this.light = level.getLightLevelDependentMagicValue(new BlockPos(this.x, this.y, this.z)) + 0.01f;
+        this.light = level.getLightLevelDependentMagicValue(new BlockPos((int)this.x, (int)this.y, (int)this.z)) + 0.01f;
         this.setColor((15f / this.light),(15f / this.light), (15f / this.light));
         this.alpha = 1;
         this.quadSize = 0.125f;
